Ensure subports status is aligned with parent port
Subports of a trunk should have the same status as the parent port.
However, with ovn, if the parent port is in DOWN status, the subports
are transitioned to ACTIVE as soon as they are attached to the trunk.

This patch ensure the status set on the subports when being attached
matches the one on the parent port.

Closes-Bug: #1957161
Change-Id: I26c43c2909b635e8b21306ea1880da3877477a17
(cherry picked from commit 15d6cfffc9f3c63a7bac75f8d04beaae82ddc794)