[OVN] Check if exists trunk ports before cleanup
While migrating from OVS to OVN one of the steps of the migration
is clean all the OVS trunk ports, this will fail if the environment
does not have any trunk ports configured.

This will do a comprovation in order to know if it's necessary
to clean them or not.

Also, since this playbook it will only clean the ovn interfaces
it is not necessary to stop the whole migration. If any error
occured while deleting any ovs interface a message will be
printed so the user can take action if necessary.

Change-Id: I6ec0b392b13daa9f64e051fb12b4b97a6c0a1730
(cherry picked from commit 9fcf42d3d524f03c8f49794bdd3730f654c146a1)