Wait longer before deleting DPDK vhu trunk bridges
In [1] we added a delay before deleting DPDK vhu trunk bridges to
mitigate a race condition when instances are rebooted. As explained in
[1], with DPDK rhu, a reboot is esentially a super fast bridge
delete-then-create that is prone to race conditions. We have recently
encountered in customer deployments that the wait added in [1] is not
long enough. As a consequence, this change increases the wait.

[1] https://review.opendev.org/c/openstack/neutron/+/717394

Change-Id: I5c1474b405d436d3b1e5db745d77999f1723b660
Partial-Bug: #1869244
(cherry picked from commit 140bb63665223d7cd2a7fee8c1d1494ebd2a802f)