[L3][QoS] Remove router QoS binding when GW is removed
Now when the external gateway of a router is removed, the QoS policy
assigned to this router is removed too. A router can only have a QoS
policy assigned if a gateway port is attached; if the port is removed,
the QoS binding must be removed too.

Closes-Bug: #1968269
Change-Id: Iced254259170972a9699aa35ae8062776e11285d
(cherry picked from commit 8ad6c3bf60081ebf19c1cadd9c46e1f5bf504892)