[OVN] Handle missing acls during log removal
During log delete we fetch all the acls for
all the pgs if there is only one log object
and do clear log of all these acls, but if
one or more acls of any of the pgs is removed
concurrently, db_set fails as acl is not found.

This patch proposes to only do log clear of those
acls which are available and add log message for the
acls which were deleted concurrently.

Also add a unit test for this case where one of
the acl get's missing.

Closes-Bug: #1971569
Change-Id: I58487024c8d0352776307f0185f0812bb00036ae
(cherry picked from commit 1471f531b86eef8928756d5e927f710a5683e6ad)